LRCI-2772 Move/modify getTeamName and helper method to BatchTestrayCaseResult
